Cleaned up all the build warnings (#223)

+++ tests/LaurelEyeEngineUnitTests/RenderTest.cpp
@@ -278,7 +278,7 @@ namespace LaurelEye {
                 renderSystem.update(dt);
             }
 #ifdef _WIN32
-            Sleep(dt * 1000.f);
+            Sleep((DWORD)(dt * 1000.f));
 #else
             sleep(dt);
 #endif
